Extract Phase 61 list validators
